Add Selector type in clicks endpoint
