+++ src/sentry/replays/lib/query.py
@@ -194,58 +194,6 @@ class Selector(Field):
             return Or(conditions)
 
 
-class WhereSelector(Field):
-    _operators = [Op.EQ, Op.NEQ]
-    _python_type = str
-
-    def as_condition(
-        self, field_alias: str, operator: Op, value: Union[List[str], str], is_wildcard: bool
-    ) -> Condition:
-        if operator == Op.NEQ:
-            return Condition(Function("identity", parameters=[1]), Op.EQ, 2)
-
-        # This list of queries implies an `OR` operation between each item in the set. To `AND`
-        # selector queries apply them separately.
-        queries: List[QueryType] = parse_selector(value)
-
-        # A valid selector will always return at least one query condition. If this did not occur
-        # then the selector was not well-formed. We return an empty resultset.
-        if len(queries) == 0:
-            return Condition(Function("identity", parameters=[1]), Op.EQ, 2)
-
-        # Conditions are pre-made and intended for application in the HAVING clause.
-        conditions: List[Condition] = []
-
-        for query in queries:
-            if query.alt:
-                conditions.append(Condition(Column("click_alt"), operator, query.alt))
-            if query.aria_label:
-                conditions.append(Condition(Column("click_aria_label"), operator, query.aria_label))
-            if query.classes:
-                conditions.append(
-                    Condition(
-                        Function("hasAll", parameters=[Column("click_class"), query.classes]),
-                        Op.EQ,
-                        1,
-                    )
-                )
-            if query.id:
-                conditions.append(Condition(Column("click_id"), operator, query.id))
-            if query.role:
-                conditions.append(Condition(Column("click_role"), operator, query.role))
-            if query.tag:
-                conditions.append(Condition(Column("click_tag"), operator, query.tag))
-            if query.testid:
-                conditions.append(Condition(Column("click_testid"), operator, query.testid))
-            if query.title:
-                conditions.append(Condition(Column("click_title"), operator, query.title))
-
-        if len(conditions) == 1:
-            return conditions[0]
-        else:
-            return And(conditions)
-
-
 class Number(Field):
     _operators = [Op.EQ, Op.NEQ, Op.GT, Op.GTE, Op.LT, Op.LTE, Op.IN, Op.NOT_IN]
     _python_type = int
